LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.measure.emHeights.worker-expected.txt


FAIL Testing emHeights assert_equals: ctx.measureText('A').emHeightAscent === 30 (got 68[number], expected 30[number]) expected 30 but got 68
